Changed age to decimal

+++ Areas/Admin/Models/DevelopmentalHistory/DevelopmentalHistoryModel.cs
@@ -5,11 +5,11 @@
         public int DevelopmentalHistoryID { get; set; }
         public int AssessmentID { get; set; }   
         public int PatientID { get; set; }
-        public int RolledOverAge { get; set; } = 0;
-        public int CrawledAge { get; set; } = 0;
-        public int WalkedAge { get; set; } = 0;
-        public int TalkedAge { get; set; } = 0;
-        public int ToiletTrainedAge { get; set; } = 0;
+        public decimal RolledOverAge { get; set; } = 0;
+        public decimal CrawledAge { get; set; } = 0;
+        public decimal WalkedAge { get; set; } = 0;
+        public decimal TalkedAge { get; set; } = 0;
+        public decimal ToiletTrainedAge { get; set; } = 0;
         public bool SpeechConcerns { get; set; } = false;
         public bool MotorSkillsConcerns { get; set; } = false;
         public bool CognitiveConcerns { get; set; } = false;
